Focusing on naive octree first
I wiped away the other implementation because it was a bit prematurely thought out, and I want to not only try a simple naive octree, but a few other simpler ideas too.

I'm still gonna need to rewrite the mesher in NaiveOctree, and probably change the order of binary cube edges to be a bit cleaner.

`subdivide_values` in `utils` now provides 8 sets of binary-ordered points, instead of a fat 27-point map. Maybe I'll make a second function to emulate the original behavior if that's desirable.