Fixed glaring issues
So, `CUBE_CORNERS` had the completely WRONG value for corner 6. AABB wasn't using this constant, but it was initializing the exact same list of points with the exact same error.

I don't know how I missed such an obvious error, but evidently there are more within because AABB is returning negative-sized AABBs (this should not happen).

Tool has been rewritten to just have one function, `value`. It's also been given a `scale` parameter, which represents the scale of the grid cube that the given point is part of. This changes the value definition from density to relative isovalue.