WORKING TERRAIN BABY
AABB was riddled with errors, so it's been rewritten and has a few unit tests just in case.

The cells were mistakenly not subdividing if the AABB totally eclipsed the cell, even if the actual isosurface was intersecting.

Tool has been rewritten. Again. Included some functions for possible concave tool support.