Experiment to make `par_generate_mesh` faster
This was an attempt to beat the performance of the single-thread `generate_mesh` function, but `par_generate_mesh` still slightly underperforms in comparison. This version of `par_generate_mesh` *is* far faster than the original, but it's pointless if doing it sequentially is the fastest.

I'm giving this commit its own branch because it has some good code and I may steal from it later on.