Back to Octants
The key system would not work as I had hoped. For every point in an octree, there is a maximum of (8*max_depth) possible cells that share it. With my key system, each point would have a maximum of 8 different keys.

It does solve corner-sharing between parent and descendants, but not among siblings. I may revisit using the idea, as this property is still useful (I wouldn't have to recurse upwards to apply modifications).

I've reverted to 58b63c637b and rewrote some parts to make a key to refer to any Octant, as any Octant in an Octree can be described by its path from the root.