apply_tool_filter
This is better than apply_tool_recursive, and is slightly faster than NaiveOctree. However, it is beat by NaiveOctree by a test in my `rayon` branch, which simply recurses in parallel.

Also, my goal of this key was to use nohash-hasher, but that was a wash. nohash is SIGNIFICANTLY slower than the default hasher, possibly due to some bucketing property. ahash, however, is slightly faster than default *and* fxhash.

This implementation of apply_tool_filter is a little sloppy. It can only iterate on the first pass, but has to individually seek out, subdivide, and reapply the tool afterwards. It also cannot collapse cells, as it cannot access the data of children during iteration.

Perhaps I can perform subdivisions in a separate HashMap, then extend them into the overall terrain.