Fixed Tools with non-contained AABBs
I wrote a note on 33a9f13 (Fixed over-subdivision, 2023-01-11) saying that Tools whose AABBs fell on the edge of a terrain wouldn't be supported, but I need to support that to enable chunking. I changed the code to first intersect the tool's AABB, and use that intersection as the tool AABB for all further tool applications. Thus, the AABB can be "inside" of the terrain and not cause any of those edge case issues. I don't know why I didn't think of this before.

I may need to fix this for OctantMap too.